Add winetricks installation and vcrun2019 setup for native ucrtbase.dll
https://forum.winehq.org/viewtopic.php?t%3D40355

The WineHQ forum thread suggests that the error
occurs because the ucrtbase.dll function
feholdexcept is not implemented in the version of
Wine being used. The recommended solutions
include:

Upgrading to a newer version of Wine where the
function might be implemented. Using a native
version of ucrtbase.dll by installing the required
Visual C++ runtime libraries using winetricks
vcrun2019. To address this issue in your
Dockerfile, we will modify it to include the
installation of the Visual C++ runtime libraries
using winetricks. This will ensure that the native
ucrtbase.dll is available. Let me make the
necessary changes.

Dockerfile-anybodycon Updated the Dockerfile to
include the installation of winetricks and the
Visual C++ runtime libraries (vcrun2019). This
should resolve the issue by providing a native
implementation of ucrtbase.dll, addressing the
unimplemented function error. Let me know if you
need further assistance!